[#4]Refactor: 태그 페이지 컴포넌트 이동
