+++ src/components/Header/Header.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import ThemeToggle from './ThemeToggle/ThemeToggle';
-import ScrollProgressBar from '../atoms/ScrollProgressBar';
+import ScrollProgressBar from './ScrollProgressBar';
 
 const HeaderBox = styled.header`
   display: flex;
