+++ src/components/Header/ScrollProgressBar.tsx
