+++ src/components/templates/Homepage.tsx
@@ -1,24 +1,27 @@
-import React from "react"
-import CategoriesWrapper from "../organisms/CategoriesWrapper";
+import React from 'react';
+import CategoriesList from '../CategoriesList/CategoriesList';
 
 type Category = {
   categoryName: string;
   quantity: number;
-}
+};
 
 type Props = {
-	children: JSX.Element,
-  categories: Category[],
+  children: JSX.Element;
+  categories: Category[];
   category?: string;
-}
+};
 
-const Homepage = ({ children, categories, category } : Props) => {
+const Homepage = ({ children, categories, category }: Props) => {
   return (
     <>
-      <CategoriesWrapper categories={categories} category={category}/>
+      <CategoriesList
+        categories={categories}
+        category={category}
+      />
       {children}
     </>
-  )
-}
+  );
+};
 
 export default Homepage;
