+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import { getAllCategories, getAllPosts, getAllTags } from '../lib/api';
 
 import PostList from '@/components/PostList/PostList';
 import Homepage from '@/components/templates/Homepage';
-import SeoHead from '@/components/SeoHead';
+import SeoHead from '@/components/SeoHead/SeoHead';
 
 import PageType from '@/types/page';
 import Post from '@/types/post';
